+++ mrmustard/lab/circuit.py
@@ -20,14 +20,12 @@ from __future__ import annotations
 
 __all__ = ["Circuit"]
 
-from abc import ABC, abstractmethod
-from typing import Dict, Iterator, Optional
+from typing import Dict, Optional
 
 from mrmustard import settings
-from mrmustard.lab.abstract import Measurement, State, Transformation
+from mrmustard.lab.abstract.operation import CircuitPart, Wire
 from mrmustard.typing import Tensor
 from mrmustard.utils.circdrawer import circuit_text
-from mrmustard.utils.tagdispenser import TagDispenser
 
 # class Circuit(Parametrized):
 #     r"""Represents a quantum circuit: a wrapper around States, Transformations and Measurements
@@ -247,252 +245,6 @@ from mrmustard.utils.tagdispenser import TagDispenser
 #         return f"< Circuit | {len(self._ops)} ops | compiled = {self._compiled} >"
 
 
-class Wire:
-    r"""A wire of a CircuitPart. It corresponds to a wire going into or coming out of an
-    object in the circuit picture. A wire corresponds to a single mode and it have one
-    or two components (called L and R) depending on the nature of the gate/state that
-    it's attached to. Wires are single when both ends are attached to pure states and/or to
-    unitary transformations, and they are double otherwise.
-
-    Arguments:
-        origin (Optional[CircuitPart]): the CircuitPart that the wire originates from
-        end (Optional[CircuitPart]): the CircuitPart that the wire is connected to
-        L (Optional[int]): the left tag of the wire (automatically generated if not specified)
-        R (Optional[int]): the right tag of the wire
-        cutoff (Optional[int]): the Fock cutoff of the wire (determined at runtime if not specified)
-    """
-
-    def __init__(
-        self,
-        origin: Optional[CircuitPart] = None,
-        end: Optional[CircuitPart] = None,
-        L: Optional[int] = None,
-        R: Optional[int] = None,
-        cutoff: Optional[int] = None,
-    ):
-        self.origin = origin
-        self.end = end
-        self.L = L or TagDispenser().get_tag()
-        self.R = R
-        self.cutoff = cutoff
-
-    @property
-    def is_connected(self) -> bool:
-        "checks if the wire is connected on both ends"
-        return self.origin is not None and self.end is not None
-
-    def connect_to(self, end: CircuitPart):
-        "connects the end of the wire to another CircuitPart"
-        if self.end:
-            raise ValueError("Wire already connected.")
-        self.end = end
-
-    @property
-    def has_dual(self):
-        "whether the wire has a dual (R) part."
-        return self.R is not None
-
-    def enable_dual(self):
-        "enables the dual (R) part of the wire."
-        if self.R:
-            raise ValueError("Wire already has dual (R) part.")
-        self.R = TagDispenser().get_tag()
-
-    def __repr__(self):
-        origin = self.origin.name if self.origin is not None else "..."
-        end = self.end.name if self.end is not None else "..."
-        tags = f"{self.L}," + (f"{self.R}" if self.R is not None else "-")
-        a = "==" if self.has_dual else "--"
-        return origin + f" {a}({tags}){a}> " + end
-
-
-class CircuitPart(ABC):
-    r"""CircuitPart supplies functionality for constructing circuits out of connected components.
-    A CircuitPart does not know about the physics of the objects it represents: it only knows about
-    its place in the circuit and how to connect to other CircuitParts.
-    """
-
-    has_dual: bool
-    input_wire_at_mode: Dict[int, Wire]
-    output_wire_at_mode: Dict[int, Wire]
-    name: str
-
-    @abstractmethod
-    def enable_dual(self) -> None:
-        "Enables the dual (R) part of all the wires throughout this CircuitPart."
-
-    @property  # NOTE: override this property in subclasses if the subclass has internal wires
-    def all_wires(self) -> Iterator[Wire]:
-        "Yields all wires of this CircuitPart (output and input)."
-        yield from (w for w in self.output_wire_at_mode.values())
-        yield from (w for w in self.input_wire_at_mode.values())
-
-    @property
-    def input_modes(self) -> set[int]:
-        "Returns the set of input modes that are used by this CircuitPart."
-        return set(self.input_wire_at_mode.keys())
-
-    @property
-    def output_modes(self) -> set[int]:
-        "Returns the set of output modes that are used by this CircuitPart."
-        return set(self.output_wire_at_mode.keys())
-
-    @property
-    def connected(self) -> bool:
-        "Returns True if at least one wire of this CircuitPart is connected to other CircuitParts."
-        return any(w.is_connected for w in self.all_wires)
-
-    @property
-    def fully_connected(self) -> bool:
-        "Returns True if all wires of this CircuitPart are connected to other CircuitParts."
-        return all(w.is_connected for w in self.all_wires)
-
-    def can_connect_output_mode(self, other: CircuitPart, mode: int) -> tuple[bool, str]:
-        r"""Checks whether this Operation can plug into another one, at a given mode.
-        This is the case if the modes exist, if they are not already connected
-        and if there is no overlap between the output and input indices.
-        In other words, the operations can be plugged as in a circuit diagram.
-
-        Arguments:
-            other (Operation): the other Operation
-            mode (int): the mode to check
-
-        Returns:
-            bool: whether the connection is possible
-        """
-        if mode not in self.output_modes:
-            return False, f"mode {mode} not an output of {self}."
-
-        if mode not in other.input_modes:
-            return False, f"mode {mode} not an input of {other}."
-
-        if self.has_dual != other.has_dual:
-            return False, "dual wires mismatch"
-
-        if self.output_wire_at_mode[mode].end is not None:
-            return (
-                False,
-                f"output of {self} already connected to {self.output_wire_at_mode[mode].end}",
-            )
-
-        if other.input_wire_at_mode[mode].origin is not None:
-            return (
-                False,
-                f"input of {other} already connected to {other.input_wire_at_mode[mode].origin}",
-            )
-
-        intersection = self.output_modes.intersection(other.input_modes)
-        input_overlap = self.input_modes.intersection(other.input_modes) - intersection
-        output_overlap = self.output_modes.intersection(other.output_modes) - intersection
-
-        if len(input_overlap) > 0:
-            return False, f"input modes overlap ({input_overlap})"
-
-        if len(output_overlap) > 0:
-            return False, f"output modes overlap ({output_overlap})"
-
-        return True, ""
-
-    def connect_output_mode(self, other: CircuitPart, mode: int) -> None:
-        "Forward-connect the wire at the given mode to the given CircuitPart."
-        can, fail_reason = self.can_connect_output_mode(other, mode)
-        if not can:
-            raise ValueError(fail_reason)
-        self.output_wire_at_mode[mode].end = other
-        # when connected the two CircuitParts share the same wire:
-        other.input_wire_at_mode[mode] = self.output_wire_at_mode[mode]
-
-
-class Operation(CircuitPart):
-    r"""A container for States, Transformations and Measurements that allows one to place them
-    inside a circuit. It contains information about which modes in the circuit the operation
-    is attached to via its wires. The Operation is an abstraction above the physics of the object
-    that it contains. Its main purpose is to allow the user to easily construct circuits."""
-
-    def __init__(
-        self,
-        wrapped: State | Transformation | Measurement,
-        input_modes: list[int],
-        output_modes: list[int],
-        has_dual: bool = False,
-    ):
-        self.wrapped = wrapped  # think of this as a fock array
-        self.name = wrapped.__class__.__qualname__
-        self.has_dual: bool = has_dual
-        self.input_wire_at_mode: Dict[int, Wire] = {
-            m: Wire(
-                end=self,
-                L=TagDispenser().get_tag(),
-                R=TagDispenser().get_tag() if has_dual else None,
-            )
-            for m in input_modes
-        }
-        self.output_wire_at_mode: Dict[int, Wire] = {
-            m: Wire(
-                origin=self,
-                L=TagDispenser().get_tag(),
-                R=TagDispenser().get_tag() if has_dual else None,
-            )
-            for m in output_modes
-        }
-
-    def __getattr__(self, name):
-        if name in self.__dict__:
-            return self.__dict__[name]
-        else:
-            return getattr(self.wrapped, name)
-
-    _repr_markdown_ = None
-
-    def disconnect(self) -> Operation:
-        "Re-issue new wires for this operation"
-        for wire_dict in [self.input_wire_at_mode, self.output_wire_at_mode]:
-            for m, wire in wire_dict.items():
-                wire_dict[m] = Wire(
-                    origin=None if wire.origin is not self else self,
-                    end=None if wire.end is not self else self,
-                    L=TagDispenser().get_tag(),
-                    R=TagDispenser().get_tag() if self.has_dual else None,
-                )
-        return self
-
-    @property
-    def modes(self) -> Optional[list[int]]:
-        "Returns the modes that this Operation is defined on"
-        if self.input_modes == self.output_modes:
-            return list(self.input_modes)
-        else:
-            raise ValueError("Operation has different input and output modes.")
-
-    def enable_dual(self) -> None:
-        "Enables the dual (R) part of all the wires of this Operation."
-        for wire in self.all_wires:
-            wire.enable_dual()
-
-    def __hash__(self):  # is this needed?
-        "hash function so that Operations can be used as keys in dictionaries."
-        tags = tuple(tag for wire in self.all_wires for tag in [wire.L, wire.R] if tag is not None)
-        return hash(tags)
-
-    def __repr__(self):
-        return (
-            f"Operation[{self.wrapped.__class__.__qualname__}](inputs={list(self.input_modes)}, "
-            + f"outputs={list(self.output_modes)}, has_dual={self.has_dual})"
-        )
-
-    def __rshift__(self, other: CircuitPart) -> Circuit:
-        other_parts = other.parts if isinstance(other, Circuit) else [other]
-        dual = self.has_dual or other.has_dual
-        if dual:
-            self.enable_dual()
-            other.enable_dual()
-        return Circuit([self] + other_parts)
-
-    def TN_tensor(self) -> Tensor:
-        "Returns the TensorNetwork Tensor of this Operation."
-        return self.wrapped.TN_tensor()
-
-
 class Circuit(CircuitPart):
     r"""A collection of interconnected Operations that can be run as a quantum device."""
 
@@ -505,12 +257,12 @@ class Circuit(CircuitPart):
         for part in self.parts:
             for mode, wire in part.input_wire_at_mode.items():
                 if not wire.is_connected:
-                    if mode in self.input_modes:
+                    if mode in self.modes_in:
                         raise ValueError("Duplicate input mode.")
                     self.input_wire_at_mode[mode] = wire
             for mode, wire in part.output_wire_at_mode.items():
                 if not wire.is_connected:
-                    if mode in self.output_modes:
+                    if mode in self.modes_out:
                         raise ValueError("Duplicate output mode.")
                     self.output_wire_at_mode[mode] = wire
         self.name = name
@@ -518,7 +270,7 @@ class Circuit(CircuitPart):
     def connect_all_parts(self):
         r"""Connects parts in the circuit according to their input and output modes."""
         for i, part1 in enumerate(self.parts):
-            for mode in part1.output_modes:
+            for mode in part1.modes_out:
                 for part2 in self.parts[i + 1 :]:
                     if part1.can_connect_output_mode(part2, mode)[0]:
                         part1.connect_output_mode(part2, mode)
@@ -538,12 +290,6 @@ class Circuit(CircuitPart):
     def flatten(self) -> Circuit:
         "Flattens the circuit."
         return Circuit([op.disconnect() for op in self.ops])
-
-    def __gt__(self, other):
-        if not isinstance(other, (Circuit, Operation)):
-            has_dual = not other.is_pure or not other.is_unitary or not other.is_projective
-            other = Operation(other, other.input_modes, other.output_modes, has_dual)
-        return Circuit([self, other])
 
     # def tag_map(self) -> dict[int, int]:
     #     tags = []
@@ -568,10 +314,10 @@ class Circuit(CircuitPart):
         "searches recursively through the circuit and returns a list of operations"
         ops = []
         for part in self.parts:
-            if isinstance(part, Operation):
-                ops.append(part)
-            elif isinstance(part, Circuit):
+            if isinstance(part, Circuit):
                 ops += part.ops
+            else:  # assuming it's an Operation now
+                ops.append(part)
         return ops
 
     def __repr__(self) -> str:
