+++ mrmustard/training/parametrized.py
@@ -48,22 +48,24 @@ class Parametrized:
     """
 
     def __init__(self, **kwargs):  # NOTE: only kwargs so that we can use the arg names
+        print(f"parametrized init called with {kwargs}")
         owner = f"{self.__class__.__qualname__}"
         self.param_names = []  # list of parameter names to preserve order
         for name, value in kwargs.items():
-            # filter out `{name}_trainable` or `{name}_bounds`` to become fields
-            # of the class as those kwargs are used to define the variables
-            if "_trainable" in name or "_bounds" in name:
+            # only keep parameters that can be trainable
+            if f"{name}_trainable" not in kwargs:
                 continue
 
             # convert into parameter class
             is_trainable = kwargs.get(f"{name}_trainable", False) or math.is_trainable(value)
             bounds = kwargs.get(f"{name}_bounds", None)
+            print(value, name, is_trainable, bounds, owner)
             param = create_parameter(value, name, is_trainable, bounds, owner)
 
             # dynamically assign parameter as attribute of the class
             self.__dict__[name] = param
             self.param_names.append(name)
+        super().__init__(self, **kwargs)
 
     def param_string(self, decimals: int) -> str:
         r"""Returns a string representation of the parameter values, separated by commas and rounded
