+++ mrmustard/utils/circdrawer.py
@@ -21,8 +21,8 @@ Operation = Any  # actually Operation, but circular import
 
 
 def mode_set(op: Operation):
-    r"""includes modes in between min and max of op.input_modes."""
-    return set(range(min(op.output_modes), max(op.output_modes) + 1))
+    r"""includes modes in between min and max of op.modes_in."""
+    return set(range(min(op.modes), max(op.modes) + 1))
 
 
 def drawable_layers(ops: list[Operation]) -> dict[int, list[Operation]]:
@@ -51,7 +51,7 @@ def add_grouping_symbols(op: Operation, layer_str: dict[int, str]) -> dict[int, 
         layer_str[min(S)] = "╭"
         layer_str[max(S)] = "╰"
         for w in range(min(S) + 1, max(S)):
-            layer_str[w] = "├" if w in op.output_modes else "│"  # other option: ┼
+            layer_str[w] = "├" if w in op.modes_out else "│"  # other option: ┼
     return layer_str
 
 
@@ -65,17 +65,17 @@ def add_op(
 
     # add the operation label and parameters
     control = []
-    if op.wrapped.__class__.__qualname__ in ["BSgate", "MZgate", "CZgate", "CXgate"]:
-        control = [list(op.output_modes)[0]]
+    if op.__class__.__qualname__ in ["BSgate", "MZgate", "CZgate", "CXgate"]:
+        control = [list(op.modes_out)[0]]
     label = op.short_name
     if decimals is not None:
         param_string = op.param_string(decimals)
         if param_string == "":
-            param_string = str(len(op.output_modes))
+            param_string = str(len(op.modes_out))
         label += "(" + param_string + ")"
 
     # add the control symbol
-    for w in op.output_modes:
+    for w in op.modes_out:
         layer_str[w] += "•" if w in control else label
 
     return layer_str
@@ -94,7 +94,7 @@ def circuit_text(
         str : String based graphic of the circuit.
     """
     # get all modes used by the ops and sort them
-    modes = sorted(list(set().union(*[op.output_modes | op.input_modes for op in ops])))
+    modes = sorted(list(set().union(*[op.modes_out | op.modes_in for op in ops])))
     # include all modes between min and max (need to draw over them)
     all_modes = range(min(modes), max(modes) + 1)
